Add `getDateRange` test
